refactor of CV support
