+++ package.json
@@ -32,7 +32,7 @@
     "@material-ui/lab": "^4.0.0-alpha.35",
     "@material-ui/styles": "^4.11.3",
     "axios": "^0.21.0",
-    "datatable-translatable": "0.11.1-3.rc-1",
+    "datatable-translatable": "0.11.1-3.rc-4",
     "deep-freeze": "^0.0.1",
     "eslint-plugin-test-selectors": "^1.3.0",
     "gitea-react-toolkit": "1.9.3",
@@ -47,7 +47,7 @@
     "scripture-resources-rcl": "4.1.6",
     "use-deep-compare-effect": "^1.3.1",
     "usfm-js": "^2.1.0",
-    "uw-content-validation": "2.1.5",
+    "uw-content-validation": "2.2.5",
     "uw-languages-rcl": "1.0.0",
     "uw-tsv-parser": "1.0.2"
   },
